requirement.txt
flask
flask-script
flask_sqlalchemy
pymysql